Cache.Store (Eiger Cache v0.1.0)

Module responsible for storing and retrieving cache data.

It uses an ETS as data store. The ETS is created with these options:

named_table to allow to reference it by its name (instead of its tid)

public to allow any process (Cache.Workers) to insert and read objects

{:read_concurrency, true} because the app aims to be used on situations where concurrent read operations are much more frequent than write operations

Functions

get(key)

@spec get(key :: any()) :: {:ok, any()} | {:error, :not_found}

It fetches and returns a value by key.

If the key doesn't exist in the table or the expires_at field is older than the current time, it returns {:error, :not_found}.

init()

@spec init() :: :ok

Creates the ETS where the cached data will be stored.

store(key, value, ttl)

@spec store(key :: any(), value :: any(), ttl :: non_neg_integer()) :: :ok

It stores a new object in the ETS.

The object is a 3-tuple where:

the 1st element is key

the 2nd element is the expiration datetime (DateTime.add(DateTime.utc_now(), ttl, :millisecond))

the 3rd element is value